added local storage with nan
