+++ layouts/Layout.tsx
@@ -8,7 +8,6 @@ const ConnectButton = dynamic(() => import('../components/ConnectButton'), { ssr
 import ScrollBar from "../components/ScrollBar";
 import Sidebar from "../components/Sidebar";
 import ShoppingCart from "../components/ShoppingCart";
-import ShopCartModal from "../components/General/ShopCartModal";
 import DarkModeButton from "../components/DarkModeButton";
 
 interface LayoutProps {
@@ -88,15 +87,11 @@ export default function Layout({ children }: LayoutProps) {
 	const parentRef = useRef<HTMLDivElement>(null)
 	const [parentDom, setParentDom] = useState<HTMLDivElement | null>(null)
 	const { address } = useAccount()
-	const [switchState, setSwitchState] = useState<boolean>(false);
-
-	const [openShopCartModal, setOpenShopCartModal] = useState<boolean>(false)
+	//const [switchState, setSwitchState] = useState<boolean>(false);
 
 	useEffect(() => {
 		setParentDom(parentRef.current)
 	}, [parentRef.current])
-
-	console.log(switchState, 'state');
 
 	return (
 		<div className="font-plus text-grey-content w-full h-screen overflow-y-scroll hidescroll" ref={parentRef}>
@@ -105,9 +100,7 @@ export default function Layout({ children }: LayoutProps) {
 			<main className="w-full min-h-screen pl-32 relative">
 				<div className={`absolute flex top-0 z-50 pr-8 right-10`}>
 					<ConnectButton />
-					<div className={`${!address ? 'hidden' : ''}`}>
-						<ShoppingCart setOpenShopCartModal={setOpenShopCartModal} />
-					</div>
+					{address && <ShoppingCart />}
 					<DarkModeButton
 					// switchState={switchState}
 					// setSwitchState={setSwitchState}
@@ -123,8 +116,6 @@ export default function Layout({ children }: LayoutProps) {
 				<Sidebar list={list} />
 			</div>
 			{parentDom && <ScrollBar parentDom={parentDom} />}
-
-			{openShopCartModal && <ShopCartModal setOpenSpecificModal={setOpenShopCartModal} />}
 		</div>
 	);
 }
